add HW_7

+++ src/services/auth.service.ts
@@ -1,7 +1,15 @@
+import { configs } from "../config/config";
+import { ActionTokenTypeEnum } from "../enums/action-token-type.enum";
 import { EmailTypeEnum } from "../enums/email-type.enum";
 import { ApiError } from "../errors/api-error";
 import { ITokenPair, ITokenPayload } from "../interfaces/token.interface";
-import { ISignIn, ISignUp, IUser } from "../interfaces/user.interface";
+import {
+  IResetPasswordSend,
+  IResetPasswordSet,
+  ISignIn,
+  IUser,
+} from "../interfaces/user.interface";
+import { actionTokenRepository } from "../repositories/action-token.repository";
 import { tokenRepository } from "../repositories/token.repository";
 import { userRepository } from "../repositories/user.repository";
 import { emailService } from "./email.service";
@@ -10,7 +18,7 @@ import { tokenService } from "./token.service";
 
 class AuthService {
   public async signUp(
-    dto: ISignUp,
+    dto: Partial<IUser>,
   ): Promise<{ user: IUser; tokens: ITokenPair }> {
     await this.isEmailExistOrThrow(dto.email);
     const password = await passwordService.hashPassword(dto.password);
@@ -25,18 +33,29 @@ class AuthService {
     const tokens = tokenService.generateTokens({
       userId: user._id,
       role: user.role,
-      deviceId: dto.deviceId,
     });
 
     await tokenRepository.create({
       ...tokens,
-      deviceId: dto.deviceId,
       _userId: user._id,
     });
 
-    // await emailService.sendMail(EmailTypeEnum.WELCOME, user.email, {
-    //   name: user.name,
-    // });
+    const token = tokenService.generateActionTokens(
+      { userId: user._id, role: user.role },
+      ActionTokenTypeEnum.VERIFY_EMAIL,
+    );
+    console.log(configs.FRONT_URL);
+    await actionTokenRepository.create({
+      type: ActionTokenTypeEnum.VERIFY_EMAIL,
+      _userId: user._id,
+      token,
+    });
+
+    await emailService.sendMail(EmailTypeEnum.WELCOME, user.email, {
+      name: user.name,
+      actionToken: token,
+      frontURL: configs.FRONT_URL,
+    });
     return { user, tokens };
   }
 
@@ -59,11 +78,9 @@ class AuthService {
     const tokens = tokenService.generateTokens({
       userId: user._id,
       role: user.role,
-      deviceId: dto.deviceId,
     });
     await tokenRepository.create({
       ...tokens,
-      deviceId: dto.deviceId,
       _userId: user._id,
     });
     return { user, tokens };
@@ -76,36 +93,82 @@ class AuthService {
     const tokens = tokenService.generateTokens({
       userId: payload.userId,
       role: payload.role,
-      deviceId: payload.deviceId,
     });
     await tokenRepository.create({
       ...tokens,
-      deviceId: payload.deviceId,
       _userId: payload.userId,
     });
-    await tokenRepository.deleteByParams({ refreshToken });
+    await tokenRepository.deleteOneByParams({ refreshToken });
     return tokens;
   }
 
-  public async logout(payload: ITokenPayload): Promise<void> {
-    const user = await userRepository.getById(payload.userId);
-    await tokenRepository.deleteByParams({
-      _userId: payload.userId,
-      deviceId: payload.deviceId,
+  public async logout(
+    jwtPayload: ITokenPayload,
+    tokenId: string,
+  ): Promise<void> {
+    const user = await userRepository.getById(jwtPayload.userId);
+    await tokenRepository.deleteOneByParams({ _id: tokenId });
+    await emailService.sendMail(EmailTypeEnum.LOGOUT, "feden2906@gmail.com", {
+      name: user.name,
     });
+  }
+
+  public async logoutAll(jwtPayload: ITokenPayload): Promise<void> {
+    const user = await userRepository.getById(jwtPayload.userId);
+    await tokenRepository.deleteManyByParams({ _userId: jwtPayload.userId });
     await emailService.sendMail(EmailTypeEnum.LOGOUT, user.email, {
       name: user.name,
     });
   }
 
-  public async logoutCompletely(payload: ITokenPayload): Promise<void> {
-    const user = await userRepository.getById(payload.userId);
-    await tokenRepository.deleteManyByParams({
-      _userId: payload.userId,
+  public async forgotPasswordSendEmail(dto: IResetPasswordSend): Promise<void> {
+    const user = await userRepository.getByEmail(dto.email);
+    if (!user) {
+      throw new ApiError("User not found", 404);
+    }
+
+    const token = tokenService.generateActionTokens(
+      { userId: user._id, role: user.role },
+      ActionTokenTypeEnum.FORGOT_PASSWORD,
+    );
+    console.log(configs.FRONT_URL);
+    await actionTokenRepository.create({
+      type: ActionTokenTypeEnum.FORGOT_PASSWORD,
+      _userId: user._id,
+      token,
     });
-    await emailService.sendMail(EmailTypeEnum.LOGOUT_COMPLETELY, user.email, {
+
+    await emailService.sendMail(EmailTypeEnum.FORGOT_PASSWORD, user.email, {
       name: user.name,
+      actionToken: token,
+      frontURL: configs.FRONT_URL,
     });
+  }
+
+  public async forgotPasswordSet(
+    dto: IResetPasswordSet,
+    jwtPayload: ITokenPayload,
+  ): Promise<void> {
+    const password = await passwordService.hashPassword(dto.password);
+    await userRepository.updateById(jwtPayload.userId, { password });
+
+    await actionTokenRepository.deleteManyByParams({
+      _userId: jwtPayload.userId,
+      type: ActionTokenTypeEnum.FORGOT_PASSWORD,
+    });
+    await tokenRepository.deleteManyByParams({ _userId: jwtPayload.userId });
+  }
+
+  public async verifyEmail(jwtPayload: ITokenPayload): Promise<void> {
+    const user = await userRepository.getById(jwtPayload.userId);
+    if (!user) {
+      throw new ApiError("User not found!", 404);
+    }
+    await actionTokenRepository.deleteManyByParams({
+      _userId: jwtPayload.userId,
+      type: ActionTokenTypeEnum.VERIFY_EMAIL,
+    });
+    await userRepository.updateById(user._id, { isVerified: true });
   }
 
   private async isEmailExistOrThrow(email: string): Promise<void> {
